% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_query.R
\name{clean_query}
\alias{clean_query}
\title{removes sql comments and prepares a human-readable query}
\usage{
clean_query(query)
}
\arguments{
\item{query}{quoted sql query text}
}
\value{
a single line of query text
}
\description{
removes sql comments and prepares a human-readable query
}
\examples{
clean_query('select * from  tbl --example comment')
}
