% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_systyp.R
\name{get_systyp}
\alias{get_systyp}
\title{Determine the type of type of database you are using (i.e. snowflake, ssms, oracle, etc)}
\usage{
get_systyp(conn)
}
\arguments{
\item{conn}{connection to database}
}
\description{
Determine the type of type of database you are using (i.e. snowflake, ssms, oracle, etc)
}
