% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_query.R
\name{read_query}
\alias{read_query}
\title{read in query if saved in a text file}
\usage{
read_query(query)
}
\arguments{
\item{query}{file path to a sql query}
}
\value{
text
}
\description{
read in query if saved in a text file
}
