% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_table.R
\name{remove_table}
\alias{remove_table}
\title{Remove a table to a database}
\usage{
remove_table(conn, tablename)
}
\arguments{
\item{conn}{connection to database}

\item{tablename}{character string name of the table in database to remove}
}
\description{
Remove a table to a database
}
