% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Save a table to a database}
\usage{
save_table(df, tablename, friendly_db_name, replace_tbl = F)
}
\arguments{
\item{df}{df to save}

\item{friendly_db_name}{specify a database name; connection parameters retrieved for db in `get_db_con()`}
}
\description{
Save a table to a database
}
