% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_message_end.R
\name{transfer_message_end}
\alias{transfer_message_end}
\title{Displays when and how long a query took to retrieve}
\usage{
transfer_message_end(dur)
}
\arguments{
\item{dur}{a proc.time() object from transfer_message_beg()}
}
\description{
Displays when and how long a query took to retrieve
}
\examples{
transfer_message_end(transfer_message_beg())
}
