% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Pass a query to a database and get a table back}
\usage{
get_data(query, friendly_db_name, cast_as_tibble = TRUE, simplify_dates = TRUE)
}
\arguments{
\item{query}{a text query in quotes ('select * from tbl') or a network path to a saved sql query (/path/saved_query.sql)}

\item{friendly_db_name}{specify a database name; connection parameters retrieved for db in `get_db_con()`}

\item{cast_as_tibble}{make the output a tibble}

\item{simplify_dates}{transform all dates in returned table from datetime to just date}
}
\value{
a table of data output specified by the query
}
\description{
Pass a query to a database and get a table back
}
