% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Save a table to a database}
\usage{
save_table(conn, df, tablename, replace_tbl = F)
}
\arguments{
\item{conn}{connection to database}

\item{df}{df to save}

\item{tablename}{character string to name the table in database}

\item{replace_tbl}{boolean to replace or append to the database table named tablename}
}
\description{
Save a table to a database
}
