% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_message_beg.R
\name{transfer_message_beg}
\alias{transfer_message_beg}
\title{Mark the beginning of a query to time the query duration}
\usage{
transfer_message_beg()
}
\value{
a proc.time() object that acts as input to transfer_message_end()
}
\description{
Mark the beginning of a query to time the query duration
}
\examples{
easydbconn::transfer_message_beg()
}
